+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [1.0.0] - 2019-12-23
 ### Added
 - path to tootown engine may be provided via env var `$TOONCHER_ENGINE_PATH`
 
@@ -43,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2017-10-31
 
-[Unreleased]: https://github.com/fphammerle/tooncher/compare/0.4.1...HEAD
+[Unreleased]: https://github.com/fphammerle/tooncher/compare/1.0.0...HEAD
+[1.0.0]: https://github.com/fphammerle/tooncher/compare/0.4.1...1.0.0
 [0.4.1]: https://github.com/fphammerle/tooncher/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/fphammerle/tooncher/compare/0.3.1...0.4.0